Update
We have edited and updated the placeholder text within the third row of the catalogue. These items had the specific, default latin text required as a template. We have removed this and changed the text to a more product-oriented caption. The about page has also been further updated with some minor fixes to the "About Us" paragraph on the about.html file.